XIVY-6202 Adapt to core changes of CompositeComponentAwareMethodExpr
In Ivy core the field compositeComponent was renamed to
compositeCompontentId and its type has changed from UIComponent to
String. This was necessary to avoid a memory leak which keeps
referencing UIComponent objects in the HTTP sessions. Causing that HTTP
sessions needs a lot more memory than usual.
Instead of the UIComponent now the id of the UIComponent has to be set
on the CompositeComponentAwareMethodExpression.